requirements.txt
Flask
firebase-admin
easyocr
Pillow
numpy
roboflow